refactor: restructure plant components architecture
