+++ src/app/(main)/page.tsx
@@ -7,8 +7,8 @@ import Input from '@/components/common/Input'
 import Fab from '@/components/common/fab'
 import SelectBox from '@/components/common/select-box'
 import { Card, CardContent } from '@/components/common/card'
-import PlantSpeciesSearchModal from '@/components/PlantSearchModal'
-import { RegisterPlantModal } from '@/components/RegisterPlantModal'
+import PlantSpeciesSearchModal from '@/components/plant/search/PlantSearchModal'
+import { RegisterPlantModal } from '@/components/plant/register/RegisterPlantModal'
 import { PerenualPlant } from '@/hooks/usePlantSearch'
 import type { PlantSpeciesInfo } from '@/types/plant'
