+++ src/components/plant/register/DateSelect.tsx
@@ -2,10 +2,10 @@
 
 import { useState } from 'react'
 import { CalendarIcon } from 'lucide-react'
-import Button from './common/button'
+import Button from '../../common/button'
 import { formatDate } from '@/utils/date'
-import { Popover } from './common/popover'
-import { Calendar } from './common/calendar'
+import { Popover } from '../../common/popover'
+import { Calendar } from '../../common/calendar'
 
 interface DateSelectProps {
   label: string
