+++ src/components/plant/search/PlantSearchInput.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import Input  from '../components/common/Input'
+import Input from '../../common/Input'
 import { Search } from 'lucide-react'
 
 interface PlantSearchInputProps {
